Apply alejandra formatting
According to the current working document for the formatting RFC
(https://github.com/nix-rfc-101/rfcs/blob/master/rfcs/0101-nix-formatting.md)
alejandra will not be used ase the base for the common formatter. But
alejandra is the formatter that produces the most similar output to the
one proposed by the RFC right now.